+++ avida-core/source/cpu/cTestCPUInterface.h
@@ -127,7 +127,7 @@ public:
   void Die(cAvidaContext& ctx) { ; } 
   void KillCellID(int target, cAvidaContext& ctx) { ; } 
   void Kaboom(int distance, cAvidaContext& ctx);
-  void Kaboom(int distance, cAvidaContext& ctx, int effect);
+  void Kaboom(int distance, cAvidaContext& ctx, double effect);
   void SpawnDeme(cAvidaContext& ctx) { ; }
   cOrgSinkMessage* NetReceive() { return NULL; } // @DMB - todo: receive message
   bool NetRemoteValidate(cAvidaContext& ctx, cOrgSinkMessage* msg) { return false; } // @DMB - todo: validate message
