+++ avida-core/source/main/cPopulationInterface.h
@@ -133,7 +133,7 @@ public:
   void Die(cAvidaContext& ctx); 
   void KillCellID(int target, cAvidaContext& ctx); 
   void Kaboom(int distance, cAvidaContext& ctx);
-  void Kaboom(int distance, cAvidaContext& ctx, int effect);
+  void Kaboom(int distance, cAvidaContext& ctx, double effect);
   void SpawnDeme(cAvidaContext& ctx); 
   int ReceiveValue();
   void SellValue(const int data, const int label, const int sell_price, const int org_id);
